+++ R/read_cifti_flat.R
@@ -9,14 +9,13 @@
 #'  included. No spatial information is included. This is the fastest way to 
 #'  read in CIFTI data. 
 #' 
-#' @inheritSection Connectome_Workbench_Description Connectome Workbench Requirement
-#' 
 #' @inheritParams cifti_fname_Param
 #' @param keep This function works by converting the CIFTI file to a GIFTI file 
 #'  and then reading it in. Should the GIFTI file be kept? If \code{FALSE}
 #'  (default), write it in a temporary directory regardless of \code{write_dir}.
 #' @param gifti_fname File path of GIFTI-format data to save the CIFTI as. 
 #'  Default: the CIFTI_fname but with the extension replaced with "flat.gii".
+#' @inheritParams idx_Param
 #' @param write_dir The directory in which to save the GIFTI, if it is being 
 #'  kept. If \code{NULL} (default), use the current working directory.
 #'
@@ -29,7 +28,7 @@
 #'
 read_cifti_flat <- function(
   cifti_fname, keep=FALSE, gifti_fname=NULL,
-  write_dir=NULL) {
+  idx=NULL, write_dir=NULL) {
 
   # ----------------------------------------------------------------------------
   # Get the output file name ---------------------------------------------------
@@ -59,6 +58,27 @@ read_cifti_flat <- function(
     sys_path(gifti_fname)
   )
   run_wb_cmd(cmd)
+
+  # Create new GIFTI with selected columns, if specified.
+  if (!is.null(idx)) {
+    stopifnot(all(idx > 0) && all(idx == round(idx)))
+    idx_is_seq <- (length(idx) > 2 && all(diff(idx) == 1))
+    if (idx_is_seq) {
+      idx_string <- paste("-column", idx[1], "-up-to", idx[length(idx)])
+    } else {
+      idx_string <- paste("-column", paste(idx, collapse=" -column "))
+    }
+    gifti_fname_original <- gifti_fname
+    gifti_fname <- file.path(
+      tempdir(), 
+      gsub("\\.gii$", ".selected_idx\\.gii", basename(gifti_fname_original))
+    )
+
+    run_wb_cmd(paste(
+      "-metric-merge", gifti_fname, "-metric", gifti_fname_original, idx_string
+    ))
+  }
+
   result <- readgii(gifti_fname)
-  result <- result$data$normal
+  do.call(cbind, result$data)
 }