+++ R/resample_surf.R
@@ -11,8 +11,12 @@
 #' 
 #' @return The new surface
 #' 
+#' @family surfing
 #' @export
 #'
+#' @section Connectome Workbench:
+#' This function interfaces with the \code{"-surface-resample"} Workbench command.
+#' 
 resample_surf <- function(
   surf, resamp_res, hemisphere=c("left", "right")){