+++ R/separate_cifti.R
@@ -19,7 +19,6 @@
 #'  medial wall vertices cannot be distinguished from one another within
 #'  \code{cortexL_fname} alone.
 #'
-#' @inheritSection Connectome_Workbench_Description Connectome Workbench Requirement
 #' 
 #' @inheritParams cifti_fname_Param
 #' @inheritParams brainstructures_Param_LR
@@ -58,8 +57,12 @@
 #' @return A named character vector with the file paths to the written 
 #'  NIFTI and GIFTI files
 #' 
+#' @family writing
 #' @export
 #'
+#' @section Connectome Workbench:
+#' This function interfaces with the \code{"-cifti-separate"} Workbench command.
+#' 
 separate_cifti <- function(cifti_fname, 
   brainstructures=c("left","right"), 
   cortexL_fname=NULL, cortexR_fname=NULL,