+++ R/view_xifti_volume.R
@@ -21,6 +21,7 @@
 #' @inheritParams verbose_Param_TRUE
 #' @param ... Additional arguments to pass to \code{papayar::papaya} or \code{oro.nifti::overlay}
 #'
+#' @family common
 #' @export
 #' @importFrom oro.nifti overlay readNIfTI as.nifti
 view_xifti_volume <- function(
@@ -76,8 +77,8 @@ view_xifti_volume <- function(
   if (verbose) {
     cat(paste0(
       "Values to be plotted range from ",
-      min(xifti$data$subcort[,idx])," to ",
-      max(xifti$data$subcort[,idx]), ".\n"
+      min(xifti$data$subcort[,idx], na.rm=TRUE)," to ",
+      max(xifti$data$subcort[,idx], na.rm=TRUE), ".\n"
     ))
   }