+++ R/write_cifti.R
@@ -143,8 +143,6 @@ write_cifti_components <- function(
 #' Write out a \code{"xifti"} object as a CIFTI file and (optionally) GIFTI 
 #'  surface files. 
 #' 
-#' @inheritSection Connectome_Workbench_Description Connectome Workbench Requirement
-#' 
 #' @inheritParams xifti_Param
 #' @inheritParams cifti_fname_Param
 #' @param surfL_fname,surfR_fname If the \[left/right\] surface is present, it 
@@ -154,8 +152,15 @@ write_cifti_components <- function(
 #'
 #' @return Named character vector of the written files
 #' 
+#' @family common
+#' @family writing
 #' @export
 #'
+#' @section Connectome Workbench:
+#' This function interfaces with the \code{"-cifti-create-dense-timeseries"},
+#'  \code{"-cifti-create-dense-scalar"}, or \code{"-cifti-create-label"} Workbench
+#'  Command, depending on the input.
+#' 
 write_cifti <- function(
   xifti, cifti_fname, surfL_fname=NULL, surfR_fname=NULL,
   verbose=TRUE) {